Move mob symbols to a separate module
