+++ lib/atom/component.js
@@ -1,6 +1,6 @@
 'use strict';
 const EventEmitter = require('events');
-const mob_symbols = require('./mob.js')._symbols;
+const mob_symbols = require('./mob/symbols.js');
 const _networked_vars = Symbol('_networked_vars');
 
 /**
